+++ onsenui/CHANGELOG.md
@@ -42,6 +42,7 @@ dev
  * ons-dialog, ons-alert-dialog, ons-popover, ons-action-sheet: Added `maskColor` property.
  * ons-alert-dialog, ons-toast, ons-modal, ons-dialog, ons-action-sheet, ons-popover: Added `animationOptions` property.
  * ons-action-sheet: Added `title` property to mirror existing `title` attribute.
+ * ons-carousel: Added `animationOptions` property.
 
  ### Misc